pyproject.toml
[tool.poetry]
name = "referral-system"
version = "0.1.0"
description = ""
authors = ["rudich <rudi4.clarinet@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
dj-database-url = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
